requirements.txt
fastapi
uvicorn
openai-whisper
pydub
transformers
torch
rutermextract